add fastspeech & test in logging

+++ tts/loss/fastspeechloss.py
@@ -15,7 +15,7 @@ class FastSpeechLoss(nn.Module):
         f = F.mse_loss
         loss_spectr = f(spectrogram_hat, spectrogram)
         loss_pitch  = f(pitch_hat, pitch)
-        loss_align  = f(alignment_hat, alignment)
+        loss_align  = f(alignment_hat.float(), alignment.float())
         loss_energy = f(energy_hat, energy)
         out = {
             "loss": loss_spectr + self.w_pitch * loss_pitch + self.w_align * loss_align +\
